CEDS-1826 Replace generic findBy method in NotificationRepository
Using all 3 parameters results in not finding any Notification, because
they contain neither EORI nor PID. It means that they can be searched by
Conversation ID only, so the safest approach is to move back to the
previous implementation.
The only difference is that now there is additional search on Submission
DB to establish if the Notifications with given Conversation ID are for
the user's Submission. In other case it means that the Notifications are
not meant to be shown for this user.