CEDS-2043 Fix parse exception for Frustrated Exports
The problem was a consequence of discrepancy in models between Backend
and Frontend services. Backend did expect all elements in Transport to
be present. This was not a problem as all 3 elements were mandatory for
'Out of the UK' exports. Introduction of 'Back into the UK' allowed not
all elements to be present and so the Backend parser was looking for
values that were not present in the JSON.