CEDS-2156 Add new request type - CreateEmptyMucr
The new request type is to cover a case where user can create an empty
MUCR by Arriving it. From our application perspective this is a separate
type of journey, so it is better to treat it this way. The output of
this operation is an Arrival request being sent to ILE.